test: refactor to use common testcfg
PR-URL: https://github.com/joyent/node/pull/25686
